requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
chromadb
pypdf
streamlit
python-dotenv
sentence-transformers
python-docx
docx2txt